[HQ-3323] Working on overhaul of the Agent: adding spring, updating the design of the Agent itself, improving code, adding generics, general code improvements and class design improvements.
Former-commit-id: 4aa60be4d8a905c589602eaad15c0c74fd60a8ce

+++ hq-agent/hq-agent-core/src/main/java/org/hyperic/hq/agent/spring/AgentApplicationContext.java
@@ -41,7 +41,7 @@ public final class AgentApplicationContext implements SmartLifecycle {
     private static final Log logger = LogFactory.getLog(AgentApplicationContext.class);
 
     private static AnnotationConfigApplicationContext ctx;
- 
+
     public static AnnotationConfigApplicationContext create(String[] basePackages, Class<?>... annotatedClasses) {
         if (ctx != null) return ctx;
   
@@ -49,8 +49,7 @@ public final class AgentApplicationContext implements SmartLifecycle {
             ctx = new AnnotationConfigApplicationContext(annotatedClasses);
             ctx.registerShutdownHook();
             ctx.scan(basePackages);
-
-
+            ctx.refresh();
         }
         catch (BeansException e) {
             shutdown();
@@ -74,13 +73,11 @@ public final class AgentApplicationContext implements SmartLifecycle {
      * Do graceful shutdown and close on demand if needed.
      */
     public static void shutdown() {
-        if (ctx != null) {
-            ctx.destroy();
-        }
+        if (ctx != null)  ctx.destroy(); 
     }
 
     public boolean isAutoStartup() {
-        return true;
+        return false;
     }
 
     public void stop(Runnable callback) {
